requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
trino
pyarrow
pyiceberg[s3fs]
